wip: fix failing tests

+++ api/passport/api.py
@@ -12,14 +12,13 @@ from ninja_extra import NinjaExtraAPI
 from ninja_extra.exceptions import APIException
 
 import api_logging as logging
+from registry.admin import get_s3_client
 from registry.api.utils import (
     aapi_key,
     check_rate_limit,
     get_analysis_api_rate_limited_msg,
     is_valid_address,
 )
-from registry.admin import get_s3_client
-from registry.api.utils import aapi_key, check_rate_limit, is_valid_address
 from registry.exceptions import InvalidAddressException
 from registry.models import BatchModelScoringRequest, BatchRequestStatus
 from scorer.settings.model_config import MODEL_AGGREGATION_NAMES
@@ -38,7 +37,7 @@ Other endpoints documented at [/docs](/docs)
 
 
 class ScoreModel(Schema):
-    score: int
+    score: float
 
 
 class DetailedScoreModel(Schema):
